security: fix LOW/INFORMATIONAL findings from security audit (#78)
L1: Move Adminer to 'debug' profile in docker-compose
- Adminer now requires explicit --profile debug flag
- Reduces attack surface in default setup

L2: Add security headers to all responses
- X-Content-Type-Options: nosniff
- X-Frame-Options: DENY
- Cache-Control: no-store
- Implemented in RequestLogger middleware

L3: Fix confusing SQL bind order in create_account
- SQL placeholders now match bind() call order
- user_id is now first in both SQL and binds

L4: Change transaction amount from INTEGER to BIGINT
- Migration: ALTER transaction.amount to BIGINT
- Updated all models/DTOs from i32 to i64
- Updated service layer and dashboard calculations
- Prevents overflow for large transaction amounts
- Matches account.balance type (already BIGINT)

All checks passing: fmt, clippy, build, tests

+++ src/database/transaction.rs
@@ -15,7 +15,7 @@ use uuid::Uuid;
 struct TransactionRow {
     id: Uuid,
     user_id: Uuid,
-    amount: i32,
+    amount: i64,
     description: String,
     occurred_at: NaiveDate,
     // Category fields
